pyproject.toml
[tool.poetry]
name = "vibe-data-director"
version = "0.1.0"
description = "FastAPI microservice for Dataset Director - Kumo SDK and HuggingFace integration"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
pandas = "^2.2.0"
pyarrow = "^17.0.0"
kumoai = "^2.6.0"
huggingface-hub = "^0.24.0"
pydantic = "^2.9.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
